Add appointme
This pull request is adding new code to an existing file, which includes changes to the AppointmentCard component and the AppointmentsPage component.

In the AppointmentCard component, the changes include updating the component to accept a new prop called onDelete, which is not used in the code. The component has also been refactored to use styled components from the Material-UI library, and some code has been added to handle appointment cancellation.

In the AppointmentsPage component, the previous hardcoded appointments data has been removed, and the state of the appointments is now set to an empty array. The component now uses the useEffect hook to fetch the appointments data from an API endpoint using the axios library. The fetched data is then used to update the state of the appointments. The AppointmentCard component is also imported into the AppointmentsPage component and is used to render each appointment.